requirements.txt
asgiref==3.5.2
black==22.3.0
click==8.1.3
Django==4.0.4
django-extensions==3.1.5
django-rest-framework==0.1.0
djangorestframework==3.13.1
Faker==13.12.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
